# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MarketPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MarketPay - A decentralized freelance marketplace built on the Stellar blockchain.\n\n## Authentication\nMost endpoints require JWT authentication obtained via the **/api/auth** challenge-response flow using Stellar wallet signatures.\n\n## Pagination\nList endpoints support cursor-based pagination via the `after` query parameter. Responses include `next_cursor` and `has_more` fields.",
    "contact": {
      "name": "Stellar MarketPay Team",
      "email": "support@stellarmarketpay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.stellarmarketpay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth (Stellar wallet signature)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT cookie set after authentication"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Developer API key for public API endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/assessments/{skill}": {
      "get": {
        "summary": "Get assessment questions for a skill",
        "tags": [
          "Assessments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "skill",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions (without answers)"
          },
          "404": {
            "description": "Unknown skill"
          }
        }
      }
    },
    "/api/assessments/{skill}/submit": {
      "post": {
        "summary": "Submit assessment answers",
        "tags": [
          "Assessments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "skill",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "answers"
                ],
                "properties": {
                  "answers": {
                    "type": "object",
                    "description": "Map of question IDs to selected option indices"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Score and result"
          },
          "429": {
            "description": "Cooldown active"
          }
        }
      }
    },
    "/api/assessments/results/{publicKey}": {
      "get": {
        "summary": "Get assessment results for a user (public)",
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "publicKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assessment results"
          }
        }
      }
    },
...
    "/api/admin/2fa/setup": {
      "post": {
        "summary": "Generate TOTP secret and QR code for admin",
        "tags": [
          "Admin 2FA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "QR code and manual entry key"
          },
          "400": {
            "description": "2FA already enabled"
          }
        }
      }
    },
    "/api/admin/2fa/verify": {
      "post": {
        "summary": "Verify TOTP, enable 2FA, upgrade JWT",
        "tags": [
          "Admin 2FA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "setup": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA enabled/verified with upgraded token"
          },
          "400": {
            "description": "Invalid code"
          }
        }
      }
    },
    "/api/admin/2fa/status": {
      "get": {
        "summary": "Get admin 2FA status",
        "tags": [
          "Admin 2FA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "2FA status including verification state"
          }
        }
      }
    }
  }
}